Assign clips into existing states and blends
This changes the typical flow of the animation import. When a controller was
found, the Importer used to just replace clips inside existing states. This
meant it wouldn't add in new animations that might have been created in the
Aseprite file. Now it iterates through all animations in the Asesprite file and adds in missing states or blend trees. It also slots clips into existing blend
trees properly.